+++ internal/api/handler/deps.go
@@ -3,11 +3,13 @@ package handler
 import (
 	"goyavision/config"
 	"goyavision/internal/adapter/mediamtx"
+	"goyavision/internal/app"
 	"goyavision/internal/port"
 )
 
 type Deps struct {
-	Repo   port.Repository
-	Cfg    *config.Config
-	MtxCli *mediamtx.Client
+	Repo              port.Repository
+	Cfg               *config.Config
+	MtxCli            *mediamtx.Client
+	WorkflowScheduler *app.WorkflowScheduler
 }